gospel_scholar_agent/requirements.txt
# Gospel Scholar Agent Dependencies

# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
supabase>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Text similarity and NLP
python-Levenshtein>=0.21.0
fuzzywuzzy>=0.18.0
nltk>=3.8.0

# API clients
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Progress and CLI
tqdm>=4.65.0
colorama>=0.4.6
click>=8.1.0

# Reporting
jinja2>=3.1.0
markdown>=3.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# HTTP client with async support
httpx>=0.24.0
aiohttp>=3.8.0
